requirements.txt
alembic==0.8.3
factory-boy==2.6.0
fake-factory==0.5.3
Flask==0.10.1
Flask-Migrate==1.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
marshmallow==2.2.1
python-editor==0.4
SQLAlchemy==1.0.9
Werkzeug==0.11.2
wheel==0.24.0
